pyproject.toml
[project]
name = "contractguard"
version = "0.1.0"
description = "8-stage autonomous contract analysis for Indian MSMEs (7 agents + autonomy loop)"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langgraph>=0.4.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "groq>=0.8.0",
    "PyMuPDF>=1.24.0",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0",
    "chromadb>=0.5.0",
    "twilio>=9.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM"]
ignore = ["S101"]  # allow assert in tests

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
